Make forecast service option to exclude for example on arm/v7 hardware
