+++ .github/workflows/build_project.yml
@@ -10,7 +10,7 @@ on:
 env:
   REGISTRY_GITHUB: ghcr.io
   IMAGE_NAME: ${{ github.repository }}
-  PACKAGE_BRANCHES: '["main", "feature-modbus-control", "master-slave-support"]'
+  PACKAGE_BRANCHES: '["main", "feature-modbus-control", "master-slave-support", "branch-cleanup"]'
 
 jobs:
   variables:
@@ -36,7 +36,6 @@ jobs:
     strategy:
       matrix:
         python-version:
-          - "3.10"
           - "3.11"
           - "3.12"
     steps:
@@ -103,8 +102,14 @@ jobs:
         with:
           fetch-depth: 0
 
-      - name: Generate static _version.py
+      - name: Install setuptools-scm and tomli
+        run: pip install setuptools-scm[toml]>=8 tomli
+
+      - name: Generate static version.json
         run: python freeze_version.py
+
+      - name: Generate requirements hashFiles
+        run: python generate_requirements.py
 
       - name: Set up QEMU
         uses: docker/setup-qemu-action@master
