+++ freeze_version.py
@@ -1,25 +1,18 @@
 import json
+from pathlib import Path
 
-import versioneer
+from setuptools_scm import get_version
 
 
-def generate_static_version_file():
-    versions = versioneer.get_versions()
-    version_string = f"""# This file was generated automatically
-import json
-    
-versions_json = '''{json.dumps(versions, indent=4)}'''
-
-def get_versions() -> dict[str | bool | None]:
-    return json.loads(versions_json)    
-
-"""
-    with open("solaredge2mqtt/_version.py", "w", encoding="utf-8") as file:
-        file.write(version_string)
-
-    print("Generated solaredge2mqtt/_version.py")
-    print(f"Version: {versions['version']}")
+def generate_version_json():
+    version = get_version(root=".", relative_to=__file__)
+    data = {"version": version}
+    path = Path("solaredge2mqtt/version.json")
+    with path.open("w", encoding="utf-8") as f:
+        json.dump(data, f, indent=2)
+        f.write("\n")  # sauberes Newline am Ende der Datei
+    print(f"Generated {path} with version {version}")
 
 
 if __name__ == "__main__":
-    generate_static_version_file()
+    generate_version_json()
