--- requirements.txt
@@ -1,16 +1,0 @@
-aiocsv==1.3.2
-aiohttp==3.11.18
-aiomqtt==2.3.2
-astral==3.2
-ephem==4.2
-influxdb-client==1.48.0
-jsonref==1.1.0
-loguru==0.7.3
-numpy==2.2.5
-pandas==2.2.3
-pydantic==2.11.4
-pyjwt==2.10.1
-pymodbus==3.9.2
-scikit-learn==1.6.1
-scipy==1.15.2
-tzlocal==5.3.1
