--- setup.cfg
@@ -1,59 +1,0 @@
-[metadata]
-name = solaredge2mqtt
-description = Read data from SolarEdge Inverter and publish it to MQTT
-long_description = file: README.md
-long_description_content_type = text/markdown
-author = Johannes Ott
-author_email = info@johannes-ott.net
-url = https://github.com/deroetzi/solaredge2mqtt
-license = MIT
-license_files = LICENSE
-classifiers =
-    Development Status :: 5 - Production/Stable
-    Topic :: Home Automation
-    License :: OSI Approved :: MIT License
-    Natural Language :: English
-    Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.11
-    Programming Language :: Python :: 3.12
-keywords = smart home
-project_urls =
-    Bug Reports = https://github.com/deroetzi/solaredge2mqtt/issues
-
-[options]
-packages = find:
-include_package_data = true
-python_requires = >=3.11, <4
-install_requires =
-    aiocsv==1.3.2
-    aiohttp==3.11.18
-    aiomqtt==2.3.2
-    astral==3.2
-    ephem==4.2
-    influxdb-client==1.48.0
-    jsonref==1.1.0
-    loguru==0.7.3
-    numpy==2.2.5
-    pandas==2.2.3
-    pydantic==2.11.4
-    pyjwt==2.10.1
-    pymodbus==3.9.2
-    scikit-learn==1.6.1
-    scipy==1.15.2
-    tzlocal==5.3.1
-
-[options.packages.find]
-exclude =
-    tests
-
-[options.entry_points]
-console_scripts =
-    solaredge2mqtt = solaredge2mqtt.service:run
-
-[versioneer]
-VCS = git
-style = pep440
-versionfile_source = solaredge2mqtt/_version.py
-versionfile_build = solaredge2mqtt/_version.py
-tag_prefix = v
-parentdir_prefix =
