--- setup.py
@@ -1,8 +1,0 @@
-import setuptools
-import versioneer
-
-if __name__ == "__main__":
-    setuptools.setup(
-        version=versioneer.get_version(),
-        cmdclass=versioneer.get_cmdclass(),
-    )
